Addressed Feedback
Thanks @afeld for all the great feedback, it made things better.

- prune README
- fix unneeded class vars
- split out ResultScanner
- simplify the main `write_summary` loop
- better temp files
- better names
- some simlification & other odds and ends
